+++ NEWS.md
@@ -1,3 +1,12 @@
+# tinyVAST 1.4.0
+
+* Adding `student` distribution, borrowing/importing code from sdmTMB
+* Changing `deviance_explained` to map off student DF at MLE to avoid negative values, in way that can be easily generalized.
+* Changing `deviance_explained` to map off overdispersion for negbin1 and negbin2
+* Changing `deviance_explained` to map of power parameter for tweedie distribution
+* add testthat checks comparing tweedie and negbin2 deviance-explained for tinyVAST with mgcv package
+* Add vignette for seasonal spatio-temporal index standardization, using Yellowtail flounder data from VAST demo
+
 # tinyVAST 1.3.0
 
 * Adding `add_vertex_covariates` to associate mesh vertices with covariates (h\t Sean Anderson)